requirements.txt
selenium==3.141.0
pytest==6.0.1
requests==2.24.0
openpyxl==3.0.5
Flask==1.1.2
jsonpath==0.82
PyMySQL==0.10.1
redis==3.5.3
BeautifulReport==0.1.3
flask_restful==0.3.8
Pillow==8.0.1
pytesseract==0.3.6
PyYAML==5.3.1
